[SBI] add tracing capability
In case that an incoming HTTP2 request contains a header field
"traceparent" copy it to the "message" struct.
In case that a "message" struct contains a "traceparent" field, send it
as a HTTP2 header field.

This preserves the traceability of requests for a particular UE.

Each HTTP2 request which has a "traceparent" field set, will be also
sent to OpenTelemetry collector, together with a "5g.ueId" field which
contains the SUPI of a traced UE.